ci(workflows): Remove upstream ruby's publish & windows workflows and slack hooks
We want to keep all the CI/test workflows for testing our patches, but
skipping the following:

1) Windows CI Workflow

It wouldn't be bad to test in Windows, but the Github Action that runs
the windows tests relies on then `msys2/setup-msys2` Github action which
is only available for the Ruby organization.

2) Publish Workflow

Needless to say, we'll be publishing Ruby in our own way.

3) Slack hooks in all workflows

Ruby had a slack hook in all CI workflows that would post a slack
message to their internal slack when a build failed. Obviously we don't
want that to happen for our builds, so that's been removed.